server/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.4
debugpy==1.8.2
decorator==5.1.1
deepdiff==7.0.0
diskcache==5.6.3
emoji==2.11.0
executing==2.0.1
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.0
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.6
install==1.3.5
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.40
llama_cpp_python==0.2.60
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mermaid-py==0.5.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
ordered-set==4.1.0
orjson==3.10.0
outcome==1.3.0.post0
packaging==23.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.8.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.11.0
unstructured==0.11.8
unstructured-client==0.22.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
